fix: history controller unit tests
